requirements.txt
# Análisis de Seguridad Colombia - Dependencias
# Python >= 3.10

# Datos
pandas>=2.0.0
openpyxl>=3.1.0      # Lectura de archivos .xlsx
xlrd>=2.0.0          # Lectura de archivos .xls (legacy)
pyarrow>=14.0.0      # Escritura Parquet
pyyaml>=6.0          # Configuración ETL

# API Dashboard
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Visualización (para fases posteriores)
# plotly>=5.18.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
